requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
gunicorn==20.1.0
Flask-JWT-Extended==4.5.2
Flask-Migrate==4.0.4